integrating error handling

+++ src/main/java/com/capgemini/training/Service/EmployeeService.java
@@ -1,6 +1,8 @@
 package com.capgemini.training.Service;
 
 import com.capgemini.training.Entity.Employee;
+import com.capgemini.training.Exception.EmployeeNotFoundException;
+import com.capgemini.training.Exception.InvalidInputException;
 import com.capgemini.training.Repository.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,6 +16,11 @@ public class EmployeeService {
     EmployeeRepository employeeRepository;
 
     public void saveEmployee(Employee emp){
+
+        if(emp.getEmpName() == null || emp.getEmpName().isEmpty()){
+            throw new InvalidInputException("Employee name cannot be empty");
+        }
+
         employeeRepository.save(emp);
     }
 
@@ -24,7 +31,7 @@ public class EmployeeService {
     public Employee updateEmployee(int id, Employee updatedEmp){
 
         Employee existingEmp = employeeRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Employee not found"));
+                .orElseThrow(() -> new EmployeeNotFoundException("Employee with ID " + id + " not found"));
 
         existingEmp.setEmpName(updatedEmp.getEmpName());
         existingEmp.setEmpEmail(updatedEmp.getEmpEmail());
@@ -36,6 +43,11 @@ public class EmployeeService {
 
 
     public void deleteEmployee(int id){
+
+        if(!employeeRepository.existsById(id)){
+            throw new EmployeeNotFoundException("Employee with ID " + id + " not found");
+        }
+
         employeeRepository.deleteById(id);
     }
 
@@ -45,7 +57,14 @@ public class EmployeeService {
     }
 
     public List<Employee> searchByName(String name){
-        return employeeRepository.findByEmpNameContainingIgnoreCase(name);
+
+        List<Employee> list = employeeRepository.findByEmpNameContainingIgnoreCase(name);
+
+        if(list.isEmpty()){
+            throw new EmployeeNotFoundException("No employees found with name: " + name);
+        }
+
+        return list;
     }
